Refactor: Unify ImageNet data loading API with CIFAR/MNIST pattern
Replace separate `imagenet_online_data()` and `toy_imagenet_data()` with
single `imagenet_data_function(root, constructor)` that returns a function
matching `(batch_size) -> (train, val, test)` signature.

Usage: `args.dataset = imagenet_data_function(path, construct_chunked_dataloaders)`

+++ compressing_classifiers_and_MLPs/src/DatasetsModels/ImageNet/imagenet_preprocessed_dataset.jl
@@ -373,8 +373,8 @@ Returns `(train, val, test)` where `test` is always `nothing`.
 """
 function imagenet_preprocessed_data(
     root::String,
-    train_batchsize,
-    val_batchsize;
+    train_batchsize::Int,
+    val_batchsize::Int;
     kwargs...
 )
     train, val = construct_chunked_dataloaders(
